pyproject.toml
[project]
name = "qq"
version = "0.1.0"
description = "qq"
readme = "README.md"
requires-python = ">=3.8, <4.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = ["memos/plugins"]
builtin_plugins = ["echo"]
 
[tool.poetry]
name = "memosbot"
version = "0.1.0"
description = ""
authors = ["Sun <sun@jiong.us>"]

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = {extras = ["fastapi"], version = "^2.3.3"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
